fix routes
