+++ apps/dotcom/client/src/routes.tsx
@@ -6,13 +6,9 @@ import {
 	SNAPSHOT_PREFIX,
 } from '@tldraw/dotcom-shared'
 import { useEffect } from 'react'
-import { Navigate, Outlet, Route, createRoutesFromElements, useRouteError } from 'react-router-dom'
+import { Outlet, Route, createRoutesFromElements, useRouteError } from 'react-router-dom'
 import { DefaultErrorFallback } from './components/DefaultErrorFallback/DefaultErrorFallback'
 import { ErrorPage } from './components/ErrorPage/ErrorPage'
-import { useApp } from './tla/hooks/useAppState'
-import { useAuth } from './tla/hooks/useAuth'
-import { useSessionState } from './tla/hooks/useSessionState'
-import { getFileUrl } from './tla/utils/urls'
 
 export const router = createRoutesFromElements(
 	<Route
@@ -60,13 +56,13 @@ export const router = createRoutesFromElements(
 		<Route path="/q/auth" lazy={() => import('./tla/pages/auth')} />
 		<Route lazy={() => import('./tla/components/TlaAppProvider')}>
 			{/* If not redirected, then local */}
-			<Route path="/q" element={<RedirectAtRoot />} />
+			<Route path="/q" lazy={() => import('./tla/components/RedirectAtRoot')} />
 			{/* Temporary file */}
 			<Route path="/q/t/:fileSlug" lazy={() => import('./tla/pages/file-temp')} />
 			{/* File view*/}
 			<Route path="/q/f/:fileSlug" lazy={() => import('./tla/pages/file')} />
 			{/* User settings */}
-			<Route path="/q/profile" element={<RequireAuthForUser />}>
+			<Route path="/q/profile" lazy={() => import('./tla/components/RequireAuthForUser')}>
 				<Route index lazy={() => import('./tla/pages/profile')} />
 			</Route>
 			{/* Internal */}
@@ -76,33 +72,3 @@ export const router = createRoutesFromElements(
 		<Route path="*" lazy={() => import('./pages/not-found')} />
 	</Route>
 )
-
-/**
- * At the workspace route, redirect to the user's most recent file.
- * Or log in and then come back here.
- */
-function RedirectAtRoot() {
-	const app = useApp()
-	const { auth, createdAt } = useSessionState()
-
-	if (!auth) throw Error('This should be wrapped in a workspace auth check')
-
-	// Navigate to the most recent file (if there is one) or else a new file
-	const file =
-		app.getUserRecentFiles(auth.userId, createdAt)?.[0]?.file ?? app.createFile(auth.userId)
-
-	return <Navigate to={getFileUrl(file.id)} replace />
-}
-
-/**
- * At the user index, an authenticated user should be taken to their workspaces.
- * The logic for what to show for a user's workspace is determined on the
- * workspaces route.
- */
-function RequireAuthForUser() {
-	const auth = useAuth()
-
-	if (!auth) throw Error('This should be wrapped in a workspace auth check')
-
-	return <Outlet />
-}
