+++ apps/dotcom/client/src/tla/components/RequireAuthForUser.tsx
@@ -1,0 +1,15 @@
+import { Outlet } from 'react-router-dom'
+import { useAuth } from '../hooks/useAuth'
+
+/**
+ * At the user index, an authenticated user should be taken to their workspaces.
+ * The logic for what to show for a user's workspace is determined on the
+ * workspaces route.
+ */
+export function Component() {
+	const auth = useAuth()
+
+	if (!auth) throw Error('This should be wrapped in a workspace auth check')
+
+	return <Outlet />
+}
